requirements.txt
anyio==3.6.2
argcomplete==2.1.1
attrs==22.2.0
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmarkgfm==0.6.0
compliance-trestle==2.0.0
cryptography==39.0.2
datamodel-code-generator==0.17.1
defusedxml==0.7.1
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
furl==2.1.3
genson==1.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ilcli==0.3.2
importlib-resources==5.12.0
inflect==5.6.2
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.3
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
openpyxl==3.1.2
orderedmultidict==1.0.1
orjson==3.8.7
packaging==23.0
paramiko==3.1.0
pathable==0.4.3
pathspec==0.11.0
platformdirs==3.1.1
prance==0.22.2.22.0
pycparser==2.21
pydantic==1.10.6
PyNaCl==1.5.0
pyrsistent==0.19.3
PySnooper==1.1.1
python-dotenv==1.0.0
python-frontmatter==1.0.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.26.15
